# server/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.0",
    "title": "Book Twitter API",
    "description": "API document for Book Twitter"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    {
      "name": "Book",
      "description": "Book API Endpoints"
    },
    {
      "name": "Booklist",
      "description": "Booklist API Endpoints"
    },
    {
      "name": "User",
      "description": "User API Endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/v1/book/": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to search books",
        "parameters": [
          {
            "name": "author",
            "in": "query",
            "type": "string"
          },
          {
            "name": "keywords",
            "in": "query",
            "type": "string"
          },
          {
            "name": "labels",
            "in": "query",
            "description": "labels this book has",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to create a book",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "authors": {
                  "example": "any"
                },
                "coverImage": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "labels": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/book/fields": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to get available fields of book",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/book/{bookId}": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to get specific book by its id",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to delete a book by its id",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/book/{bookId}/review": {
      "post": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to create a book review",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rank": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/book/{bookId}/review/{username}": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to get book review of specific user",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/booklist/": {
      "get": {
        "tags": [
          "Booklist"
        ],
        "description": "Endpoint to search booklists",
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "type": "string"
          },
          {
            "name": "labels",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Booklist"
        ],
        "description": "Endpoint to create a booklist",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "books": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "labels": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "public": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/booklist/fields": {
      "get": {
        "tags": [
          "Booklist"
        ],
        "description": "Endpoint to get available fields of booklist",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/booklist/{booklistId}": {
      "get": {
        "tags": [
          "Booklist"
        ],
        "description": "Endpoint to get specific booklist by its id",
        "parameters": [
          {
            "name": "booklistId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Booklist"
        ],
        "description": "Endpoint to delete a booklist by its id",
        "parameters": [
          {
            "name": "booklistId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/booklist/{booklistId}/review": {
      "post": {
        "tags": [
          "Booklist"
        ],
        "description": "Endpoint to create a booklist review",
        "parameters": [
          {
            "name": "booklistId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rank": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/booklist/{booklistId}/review/{username}": {
      "get": {
        "tags": [
          "Booklist"
        ],
        "description": "Endpoint to get booklist review of specific user",
        "parameters": [
          {
            "name": "booklistId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/user/{username}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get user info by its username",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/user/": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to post user info",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "avatar": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "hashedPass": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/user/{userId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to delete user info by its id",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}